Add skeleton for frame layout parser
