+++ src/frameparser.cpp
@@ -1,0 +1,5 @@
+#include "frameparser.h"
+
+FrameParser::FrameParser(std::string buf) {
+}
+
